Bam MM parsing and VCF / BCF support
In the middle of adding support for:

1. Parsing MM and ML auxiliary tags in bam / sam files. These are
somewhat complex to parse and this functionality is restricted to
forward aligning reads and does not yet infer the reference positions
for modifications.
2. Parsing vcf / bcf files into R data. Again, this is not yet complete
and there are many important things missing.

For both of these, memory errors and other bugs are possible. In
particular the wrapper functions need to be rewritten as the structure
of the returned data has changed.