Bam MM and ML parsing
MM and ML aux tabs appear to be parsed correctly for alignments that:

1. Do no have the 0x10 flag set (i.e. not reversed)
2. Are not hard clipped.

The interface for the sam_read_n() function has changed such that the
returned R data structure is different. This should be handled correctly
by the R wrapper function provided.

The functions should be considered experimental; there are still likely
to be bugs hidden in the code for some time.