Quality values for mismatched bases
The alignments_region() function has been modified to
include the quality values in the 'nuc' position of
the 'diff' component of the list returned. The 'diff'
component is a matrix giving the reference and query
positions of mismatched residues. This matrix also
contains a 'nuc' column that encodes the reference
and query nucleotides in the lower 16 bits of a single
integer. These values now also include the quality of
the query nucleotides in the third byte (bits 16-23).

Also modified a printf statement to remove warnings.