Bug squash: incorrect use of constant
An error was introduced as part of a cleanup remplacing literal numbers
with pre-processor constants. The flag AR_CIG was used instead of AR_Q_DEPTH
when initialising memory; AR_Q_DEPTH was then checked later before
leading to an attempt to use unallocated memory leading to a
segmentation fault.