Incorrect sequencing depth offset fixed
The offset for sequence depth was calculated based on an alignment
start position that had been converted from being 0 to 1-based. This
resulted in the sequencing depth being shifted one to the right (i.e.
the depth reported at position n, was that for position n-1).

This is now fixed.