Query to reference alignment function
Added a function that exports the query_to_ref function to R.
This makes it possible to map query coordinates (obtained
for example by aligning cDNA sequences) to their genome
positions based on their alignments to the genome.

As part of this, I also added two extra columns to the
al table returned by aligned.region; these columns give
the range of the rows in the cigar ops table.