Refactored alignments_region() to use cigar_to_table()
The parsing of cigar data was moved out of alignments_region()
and into cigar_to_table().

This means that both sam_read_n() and alignments_region() parse
cigar operations externally reducing the amount of code redundancy.

This should also make it easier to use pthreads to increase the
parsing speed.