Multithreaded regions_alignment
alignment_regions_mt()

now seems to work properly. However, the speedup for parsing alignments
is modest; up to 5 times for around 20 threads. The increase in
throughput seems to reach a maxium at around 20 threads. This may
be due either to the additonal single threaded work that has to be done
(copying strings), or to limitations in I/O.

The code has not been checked with Valgrind.

The writing of this part, strongly made the argument for changing data
structures from imatrix to kvecs of defined structs. That should make
it much easier to avoid the many mistakes made when making this work.