alignments_region_mt() Valgrind check
Valgrind detected a small number of errors due to failure to
clean up htslib resources. These have now been checked and
Valgrind gives a clean result.

-Wall still gives warnings about unused 'static const' variables
in common.h. These _are_ used in files that include common.h.
Not sure how to get rid of those warnings, but it would be nice to
do so in order to declutter.