warning() calls in threaded code removed
Calls to R's warning() function in code that can be called by threaded
codes were removed. These caused stack errors. This behaviour is
consistent with R documentation and the lack of thread-safety for
calls to R internal functions. The calls were present due to reuse of
code that had previously only been called from non-threaded functions.

It may be an idea to either switch these warning statements to printf()
or to make them optional by the caller as they have previously been
useful in identifying errors in the code.

The calculation of query start coordinates and query sequence offsets
was also split as these will be different when cigar data contains
hard clip operations. This was causing issues for identifying
differences between query and reference sequences.