Auxiliary integer fields
Added a function to extract standard integer auxiliary fields. This
is currently called by alignments_region and alignments_region_thread
with 3 common (and one value I'd like to be common) auxiliary fields
that are added to the core alignment table. It would be better for
the user to specify the tags to be extracted, but this will require a
non-fixed return data structure. This could be done in a similar way
to how I parse auxiliary strings at the moment.

The order of arguments to aligned.region() was also changed to make it
more convenient to call it. In particular the name of the index file has
been moved to the end since this isn't usually specified.