Squashed aux field regression and expected error estimate
The parsing of a set of standard auxiliary fields by alignments_region()
introduced a bug in the multithreaded merge function. The merge function
made a (bad) assumption that the ops.0 and ops.1 fields were the last
fields of the alignment table. Adding auxiliary fields to this upset
this (stupid / rushed) assumption. I have introduced a new constant to
handle this (AR_AL_OPS_ROW) and the threaded and non-threaded versions
now return identical data.

I've also introduced a function that estimates the expected number of
errors for a given query sequence. This is based on a Poisson
approximation of the Poisson Binomial distribution of the quality
values. In essence is it simply, sum( pow(10, q/-10) ). The reported
value is multiplied by 256 to allow for some non-integral
approximations. This number can be compared to the edit distance; though
the comparison only makes sense if the complete sequence has been
aligned.