added: unique_query_pairs, wrapped by uniq.query.pairs
Added a function that identifies unique read pairs from a set
of user defined regions. This allows for semi-unique mappings
where clear homology exists between regions (contigs, scaffolds,
chromosomes). That is, a read can map to two different regions if
they are considered homologous and not be considered as non-unique.

The function has not been fully tested and does not at the moment
consider the quality of the alignment because it relies on the
query name hash function. This does not harvest any of the auxiliary
data fields and hence only has the mapping quality to consider. If
a read has been mapped to two identical regions, then this will be 0
for both alignments and so does not provide sufficient information.