Additional hiC read pair extraction
Added a more complex read pair extraction function that should be
more suitable for sequences mapped to a haplotype resolved assembly.

Most reads mapped to a haplotype resolved assembly (i.e. mapping to
both haplotypes at the same time) will map equally well to both
haplotypes. This means that they will have a mapping quality of 0, and
that mapping quality cannot be used in the same manner. The new
function, extract_read_pairs_2 tries to compensate for this by using
alignment scores.

The new function allows certain scaffolds to be treated specially.
The idea being that scaffolds that contain large numbers of repeated
sequences will have no or very few unique alignments; however if all are
to the same scaffold, and the other pair maps uniquely to a
non-repetitive scaffold then this should still be informative.

This function is still being tested as I'm observing some strange
behaviour that I'm unable to explain.